+++ src/index.js
@@ -3,12 +3,11 @@ import '../styles/main.scss';
 export * from '@opuscapita/react-alerts';
 export * from '@opuscapita/react-cards';
 export * from '@opuscapita/react-icons';
-
-export * from './dropdown-container/index.js';
-export * from './dropdown-menu/index.js';
+export * from '@opuscapita/react-spinner';
+export * from '@opuscapita/react-dropdown-container';
+export * from '@opuscapita/react-dropdown-menu';
 export * from './dropdown-multi-select/index.js';
 export * from './responsive-navbar/index.js';
-export * from './spinner/index.js';
 export * from './split-pane/index.js';
 export * from './layouts/application/index.js';
 export * from './extended-search/index.js';